# Optimize AspNetCoreDiagnosticObserver path