--- a/src/Datadog.Trace/Sampling/DefaultSamplingRule.cs
+++ b/src/Datadog.Trace/Sampling/DefaultSamplingRule.cs
@@ -8,7 +8,7 @@ namespace Datadog.Trace.Sampling
     {
         private static readonly Vendors.Serilog.ILogger Log = DatadogLogging.For<DefaultSamplingRule>();
 
-        private static Dictionary<string, float> _sampleRates = new Dictionary<string, float>();
+        private static Dictionary<SampleRateKey, float> _sampleRates = new Dictionary<SampleRateKey, float>();
 
         public string RuleName => "default-rule";
 
@@ -16,6 +16,8 @@ namespace Datadog.Trace.Sampling
         /// Gets the lowest possible priority
         /// </summary>
         public int Priority => int.MinValue;
+
+        internal static bool UseOptim { get; set; }
 
         public bool IsMatch(Span span)
         {
@@ -34,7 +36,7 @@ namespace Datadog.Trace.Sampling
             var env = span.GetTag(Tags.Env);
             var service = span.ServiceName;
 
-            var key = $"service:{service},env:{env}";
+            var key = new SampleRateKey(service, env);
 
             if (_sampleRates.TryGetValue(key, out var sampleRate))
             {
@@ -51,21 +53,92 @@ namespace Datadog.Trace.Sampling
         {
             // to avoid locking if writers and readers can access the dictionary at the same time,
             // build the new dictionary first, then replace the old one
-            var rates = new Dictionary<string, float>(StringComparer.OrdinalIgnoreCase);
+            var rates = new Dictionary<SampleRateKey, float>();
 
             if (sampleRates != null)
             {
                 foreach (var pair in sampleRates)
                 {
                     // No point in adding default rates
-                    if (pair.Value != 1)
+                    if (pair.Value == 1.0f)
                     {
-                        rates.Add(pair.Key, pair.Value);
+                        continue;
                     }
+
+                    var key = SampleRateKey.Parse(pair.Key);
+
+                    if (key == null)
+                    {
+                        Log.Warning("Could not parse sample rate key {0}", pair.Key);
+                        continue;
+                    }
+
+                    rates.Add(key.Value, pair.Value);
                 }
             }
 
             _sampleRates = rates;
+        }
+
+        private readonly struct SampleRateKey : IEquatable<SampleRateKey>
+        {
+            private static readonly char[] PartSeparator = new[] { ',' };
+            private static readonly char[] ValueSeparator = new[] { ':' };
+
+            private readonly string _service;
+            private readonly string _env;
+
+            public SampleRateKey(string service, string env)
+            {
+                _service = service;
+                _env = env;
+            }
+
+            public static SampleRateKey? Parse(string key)
+            {
+                // Expected format:
+                // service:{service},env:{env}
+                var parts = key.Split(PartSeparator);
+
+                if (parts.Length != 2)
+                {
+                    return null;
+                }
+
+                var serviceParts = parts[0].Split(ValueSeparator, 2);
+
+                if (serviceParts.Length != 2)
+                {
+                    return null;
+                }
+
+                var envParts = parts[1].Split(ValueSeparator, 2);
+
+                if (envParts.Length != 2)
+                {
+                    return null;
+                }
+
+                return new SampleRateKey(serviceParts[1], envParts[1]);
+            }
+
+            public bool Equals(SampleRateKey other)
+            {
+                return _service == other._service && _env == other._env;
+            }
+
+            public override bool Equals(object obj)
+            {
+                return obj is SampleRateKey other && Equals(other);
+            }
+
+            public override int GetHashCode()
+            {
+                unchecked
+                {
+                    return ((_service != null ? _service.GetHashCode() : 0) * 397) ^ (_env != null ? _env.GetHashCode() : 0);
+                }
+            }
         }
     }
 }