- Do not call `ExtractHeaderTags` when no header tag has been set
- Use an enumerator instead of a dictionary in `ExtractHeaderTags`
- Add an overload to `GetRelativeUrl` to avoid useless Uri creation
- Cache the string representation of status code 200 (note: this optimization can probably be applied to other code paths, and possibly be extended to other common status codes)

Benchmark: calling `AspNetCoreDiagnosticObserver.OnHostingHttpRequestInStart`:

With no header tags set:

```
|            Method |     Mean |     Error |    StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------ |---------:|----------:|----------:|------:|--------:|-------:|-------:|------:|----------:|
| RequestStart_Old  | 4.196 us | 0.4813 us | 0.3183 us |  1.00 |    0.00 | 0.3433 | 0.1068 |     - |   2.12 KB |
| RequestStart_New  | 4.098 us | 0.5276 us | 0.3490 us |  0.98 |    0.07 | 0.2365 | 0.0763 |     - |   1.46 KB |

```

With header tags:

```
|            Method |     Mean |     Error |    StdDev | Ratio | RatioSD |  Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------------ |---------:|----------:|----------:|------:|--------:|-------:|-------:|------:|----------:|
| RequestStart_Old  | 5.406 us | 0.5658 us | 0.3743 us |  1.00 |    0.00 | 0.4730 | 0.1373 |     - |   2.94 KB |
| RequestStart_New  | 4.642 us | 0.5638 us | 0.3729 us |  0.86 |    0.02 | 0.3662 | 0.1068 |     - |   2.28 KB |

```

(note: the benchmarks include the optimization done in https://github.com/DataDog/dd-trace-dotnet/pull/902)